+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 ### Changes
-- Add rendering for jam in the vat.
-- Fix a dupe.
+- Smoke up to 8 items per string.
+- Changed stomping barrel recipe.
+- Allowed melting copper pipes.
+- Fix placing jute fiber allowing the deletion of blocks.